faiss-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
faiss-cpu==1.9.0
numpy==1.26.4
